requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9